Moved Search component into homes page
